pyproject.toml
[tool.poetry]
name = "caiper"
version = "0.1.0"
description = ""
authors = ["David Sonder <dave.sonder@protonmail.ch>"]

[tool.poetry.dependencies]
python = "^3.9"
pycairo = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
